+++ packages/angular/CHANGELOG.md
@@ -1,3 +1,39 @@
+# v0.6.0 (Tue Oct 17 2023)
+
+#### 🚀 Enhancement
+
+- feat: Add invalid and invalidMessage to File Input and remove layout bottom margin [#121](https://github.com/UKHO/admiralty-design-system/pull/121) ([@UKHOGreg](https://github.com/UKHOGreg))
+
+#### Authors: 1
+
+- Greg Williams ([@UKHOGreg](https://github.com/UKHOGreg))
+
+---
+
+# v0.5.0 (Wed Sep 13 2023)
+
+#### 🚀 Enhancement
+
+- fix: Input Invalid consistency [#110](https://github.com/UKHO/admiralty-design-system/pull/110) ([@UKHOGreg](https://github.com/UKHOGreg))
+
+#### Authors: 1
+
+- Greg Williams ([@UKHOGreg](https://github.com/UKHOGreg))
+
+---
+
+# v0.4.6 (Thu Aug 31 2023)
+
+#### 🐛 Bug Fix
+
+- Adding in invalid props to the radio group [#107](https://github.com/UKHO/admiralty-design-system/pull/107) ([@ecstaticrainbow](https://github.com/ecstaticrainbow))
+
+#### Authors: 1
+
+- Tom Barham ([@ecstaticrainbow](https://github.com/ecstaticrainbow))
+
+---
+
 # v0.4.4 (Thu Aug 24 2023)
 
 #### 🐛 Bug Fix